# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Komm — Bulk Messaging Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, scheduled, sending, sent, failed]
        - name: channel
          in: query
          schema:
            type: string
            enum: [sms, whatsapp, email]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignsPage"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign with delivery stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
    put:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign (draft only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/send:
    post:
      operationId: sendCampaign
      tags: [campaigns]
      summary: Send a campaign now (or confirm schedule)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/messages:
    get:
      operationId: listCampaignMessages
      tags: [campaigns]
      summary: List individual message delivery records for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, sent, delivered, failed, opened]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Message delivery records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignMessagesPage"

  # ── Inbox ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Campaign:
      type: object
      required: [id, name, channel, status, recipientCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        channel:
          type: string
          enum: [sms, whatsapp, email]
        status:
          type: string
          enum: [draft, scheduled, sending, sent, failed]
        body:
          type: string
        templateId:
          type: integer
          nullable: true
        groupIds:
          type: array
          items:
            type: integer
        recipientCount:
          type: integer
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    CampaignDetail:
      allOf:
        - $ref: "#/components/schemas/Campaign"
        - type: object
          properties:
            deliveredCount:
              type: integer
            failedCount:
              type: integer
            openedCount:
              type: integer
            deliveryRate:
              type: number
    CreateCampaignBody:
      type: object
      required: [name, channel, body]
      properties:
        name:
          type: string
        channel:
          type: string
          enum: [sms, whatsapp, email]
        body:
          type: string
        templateId:
          type: integer
          nullable: true
        groupIds:
          type: array
          items:
            type: integer
        scheduledAt:
          type: string
          format: date-time
          nullable: true
    CampaignsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Campaign"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CampaignMessage:
      type: object
      required: [id, campaignId, contactId, contactName, phone, status, createdAt]
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        contactId:
          type: integer
        contactName:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [pending, sent, delivered, failed, opened]
        errorMessage:
          type: string
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    CampaignMessagesPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CampaignMessage"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Inbox ────────────────────────────────────────────────────────────────
...