# lib/api-zod/src/generated/types/campaignMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Komm — Bulk Messaging Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignMessageStatus } from "./campaignMessageStatus";

export interface CampaignMessage {
  id: number;
  campaignId: number;
  contactId: number;
  contactName: string;
  phone: string;
  status: CampaignMessageStatus;
  errorMessage?: string | null;
  deliveredAt?: Date | null;
  createdAt: Date;
}