# lib/api-zod/src/generated/types/campaignsPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Komm — Bulk Messaging Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Campaign } from "./campaign";

export interface CampaignsPage {
  data: Campaign[];
  total: number;
  page: number;
  limit: number;
}